Split existing views into separate frontend / admin modules
