+++ etna/feedback/views/__init__.py
@@ -1,0 +1,2 @@
+from .admin import *  # noqa
+from .frontend import *  # noqa
