+++ etna/feedback/views/frontend.py
@@ -18,24 +18,24 @@ from django.shortcuts import get_object_or_404
 from django.urls import reverse
 from django.utils.decorators import method_decorator
 from django.utils.http import urlencode
-from django.utils.translation import gettext_lazy as _
 from django.views.decorators.csrf import csrf_exempt
 from django.views.generic import FormView, TemplateView
 
-from wagtail.admin.auth import permission_denied
-from wagtail.admin.filters import WagtailFilterSet
-from wagtail.admin.views.reports import ReportView
-from wagtail.admin.widgets import AdminDateInput
 from wagtail.models import Revision
 
-import django_filters
-
-from etna.feedback.constants import SentimentChoices
 from etna.feedback.forms import FeedbackCommentForm, FeedbackForm
 from etna.feedback.models import FeedbackPrompt, FeedbackSubmission
 from etna.feedback.utils import sign_submission_id
 
 logger = logging.getLogger(__name__)
+
+
+__all__ = [
+    "FeedbackSubmitView",
+    "FeedbackSuccessView",
+    "FeedbackCommentSubmitView",
+    "FeedbackCommentSuccessView",
+]
 
 
 class VersionedFeedbackViewMixin:
@@ -215,75 +215,3 @@ class FeedbackCommentSuccessView(VersionedFeedbackViewMixin, TemplateView):
             next_url=self.request.GET.get("next", "/"),
             **kwargs,
         )
-
-
-class FeedbackSubmissionFilterSet(WagtailFilterSet):
-    received_from = django_filters.DateFilter(
-        field_name="received_at",
-        lookup_expr="date__gte",
-        label=_("Date received (from)"),
-        widget=AdminDateInput,
-    )
-    received_to = django_filters.DateFilter(
-        field_name="received_at",
-        lookup_expr="date__lte",
-        label=_("Date received (to)"),
-        widget=AdminDateInput,
-    )
-    response_sentiment = django_filters.ChoiceFilter(
-        choices=SentimentChoices.choices, label=_("Sentiment")
-    )
-
-    class Meta:
-        model = FeedbackSubmission
-        fields = {
-            "path": ["iexact", "istartswith"],
-        }
-
-
-class FeedbackSubmissionReportView(ReportView):
-    title = "Feedback submissions"
-    header_icon = "form"
-    model = FeedbackSubmission
-    is_searchable = False
-    filterset_class = FeedbackSubmissionFilterSet
-    template_name = "feedback/reports/submission_report.html"
-
-    list_display = [
-        "received_at",
-        "path",
-        "prompt_text",
-        "response",
-        "comment_truncated",
-    ]
-    list_export = [
-        "id",
-        "public_id",
-        "received_at",
-        "full_url",
-        "path",
-        "query_params",
-        "prompt_text",
-        "response_label",
-        "response_sentiment",
-        "sentiment_label",
-        "comment_prompt_text",
-        "comment",
-        "prompt_id",
-        "prompt_revision_id",
-        "page_id",
-        "page_revision_id",
-        "page_revision_published",
-        "user_id",
-        "site_id",
-    ]
-
-    def dispatch(self, request, *args, **kwargs):
-        if not self.request.user.is_superuser:
-            return permission_denied(request)
-        return super().dispatch(request, *args, **kwargs)
-
-    def get_context_data(self, *args, **kwargs):
-        context = super().get_context_data(*args, **kwargs)
-        context["table"].base_url = self.request.path
-        return context
